Some fixes

+++ app/controllers/post.controller.js
@@ -4,7 +4,8 @@ const notifCtrl = require("../controllers/notification.controller");
 const fs = require("fs");
 
 // Set image path and make folder
-const imagePath = "./public/images/post/";
+const prefixPath = "images/post";
+const imagePath = "./public/" + prefixPath + "/";
 if (!fs.existsSync(imagePath)) {
   fs.mkdirSync(imagePath, { recursive: true });
 }
@@ -56,9 +57,19 @@ exports.readAll = async (req, res) => {
     if (community == null)
       throw new Error("La communauté spécifié est introuvable.");
 
-    let posts = await community.getPosts();
+    let posts = await community.getPosts({
+      include: [db.PostFile, db.PostComment, db.PostLike],
+    });
     if (posts.length == 0)
       throw new Error("Il n'y a aucun poste dans cette communauté.");
+
+    // Set image full url
+    const baseUri = req.protocol + "://" + req.get("host");
+    posts.forEach((post) => {
+      post.PostFiles.forEach((file) => {
+        file.file = baseUri + "/" + prefixPath + "/" + file.file;
+      });
+    });
 
     return Helper.successResponse(req, res, { posts }, hateoas(req));
   } catch (error) {
@@ -75,8 +86,16 @@ exports.readAll = async (req, res) => {
  */
 exports.readOne = async (req, res) => {
   try {
-    let post = await db.Post.findByPk(req.params.postId);
+    let post = await db.Post.findByPk(req.params.postId, {
+      include: [db.PostFile, db.PostComment, db.PostLike],
+    });
     if (post == null) throw new Error("Ce poste n'existe pas.");
+
+    // Set image full url
+    const baseUri = req.protocol + "://" + req.get("host");
+    post.PostFiles.forEach((file) => {
+      file.file = baseUri + "/" + prefixPath + "/" + file.file;
+    });
 
     return Helper.successResponse(req, res, { post }, hateoas(req));
   } catch (error) {
@@ -223,8 +242,7 @@ exports.delete = async (req, res) => {
     let files = post.getPostFiles();
     files.forEach((file) => {
       // Delete image
-      if(fs.existsSync(imagePath / file))
-        fs.unlinkSync(imagePath + file);
+      if (fs.existsSync(imagePath + file)) fs.unlinkSync(imagePath + file);
 
       // Delete in db
       file.destroy();
@@ -246,20 +264,20 @@ exports.delete = async (req, res) => {
  * @param {*} res
  * @returns response
  */
- exports.readFiles = async (req, res) => {
+exports.readFiles = async (req, res) => {
   try {
-
     let post = await db.Post.findByPk(req.params.postId);
     if (post == null) throw new Error("Ce poste n'existe pas.");
 
     // Get image files
-    let images = post.getPostFiles();
+    let files = await post.getPostFiles();
     const baseUri = req.protocol + "://" + req.get("host");
-    Array.prototype.forEach((image, index, images) => {
-      image.file = baseUri + "/" + imagePath + image.file;
+
+    files.forEach(image => {
+      image.file = baseUri + "/" + prefixPath + "/" + image.file;
     });
 
-    return Helper.successResponse(req, res, { images }, hateoas(req));
+    return Helper.successResponse(req, res, { files }, hateoas(req));
   } catch (error) {
     console.error(error);
     return Helper.errorResponse(req, res, error.message);
@@ -343,6 +361,12 @@ function hateoas(req) {
       href: baseUri + "/api/post",
     },
     {
+      rel: "upload",
+      method: "POST",
+      title: 'Upload a file to a post',
+      href: baseUri + "/api/post" + (req.params.postId || ":postId") + "/file",
+    },
+    {
       rel: "readAll",
       method: "GET",
       title: "List all Community Posts",
@@ -356,6 +380,12 @@ function hateoas(req) {
       method: "GET",
       title: "Read one Post",
       href: baseUri + "/api/post/" + (req.params.postId || ":postId"),
+    },
+    {
+      rel: "readFiles",
+      method: "GET",
+      title: "Read all files from this post",
+      href: baseUri + "/api/post/" + (req.params.postId || ":postId") + "/files",
     },
     {
       rel: "like",
@@ -375,14 +405,21 @@ function hateoas(req) {
       method: "PUT",
       title: "Update a Post",
       href:
-        baseUri + "/api/post/" + (req.params.postId || ":postId") + "/update",
+        baseUri + "/api/post/" + (req.params.postId || ":postId") + "",
     },
     {
       rel: "delete",
       method: "DELETE",
       title: "Delete a Post",
       href:
-        baseUri + "/api/post/" + (req.params.postId || ":postId") + "/delete",
+        baseUri + "/api/post/" + (req.params.postId || ":postId") + "",
+    },
+    {
+      rel: "deleteFile",
+      method: "DELETE",
+      title: "Delete a File from this post",
+      href:
+        baseUri + "/api/post/" + (req.params.postId || ":postId") + "/file",
     },
   ];
 }
